@@ -3,61 +3,72 @@
   "version": "0.0.0",
   "description": "kotlinlang.org site",
   "license": "Apache-2.0",
+  "browserslist": [
+    "defaults, not ie > 0, edge > 18"
+  ],
   "dependencies": {
     "@jetbrains/babel-preset-jetbrains": "^1.0.1",
-    "advanced-url-loader": "0.0.3",
+    "@jetbrains/kotlin-web-site-ui": "^2.1.15",
+    "@rescui/typography": "0.0.1",
+    "@rescui/ui-contexts": "0.0.2",
+    "@rollup/plugin-buble": "0.21.3",
+    "algoliasearch": "4.10.2",
     "aos": "2.1.1",
     "autoprefixer": "6.3.1",
     "babel-core": "6.18.2",
-    "babel-loader": "^7.1.1",
+    "babel-loader": "7",
     "babel-plugin-transform-runtime": "^6.23.0",
-    "clean-webpack-plugin": "0.1.18",
-    "codemirror": "5.39.0",
+    "codemirror": "5.61.0",
     "core-js": "^2.4.1",
-    "cross-env": "1.0.7",
-    "css-loader": "0.21.0",
+    "css-loader": "5.2.6",
+    "csso-webpack-plugin": "2.0.0-beta.1",
     "debounce": "1.0.2",
     "event-emitter": "0.3.4",
-    "exports-loader": "^0.6.3",
-    "extend": "2.0.1",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "0.9.0",
+    "extend": "3.0.2",
+    "file-loader": "6.2.0",
     "fixer.js": "1.4.15",
-    "highlight.js": "9.13.1",
-    "imports-loader": "^0.7.0",
+    "highlight.js": "10.5.0",
     "in-viewport": "3.4.2",
-    "instantsearch.js": "2.0.2",
-    "jquery": "3.0.0",
-    "kotlin-playground": "1.16.0",
-    "loader-utils": "1.1.0",
-    "mustache-loader": "1.0.0",
-    "node-sass": "^4.5.3",
-    "nunjucks": "^3.0.1",
-    "nunjucks-loader": "^2.4.8",
-    "postcss-loader": "^2.0.6",
+    "instantsearch.js": "4.24.1",
+    "jquery": "3.5.0",
+    "kotlin-playground": "1.22.0",
+    "mini-css-extract-plugin": "1.6.1",
+    "mini-svg-data-uri": "1.1.3",
+    "mustache-loader": "1.4.3",
+    "node-sass": "6.0.1",
+    "nunjucks": "3.2.2",
+    "nunjucks-loader": "3.0.0",
+    "postcss-custom-properties": "10.0.0",
+    "postcss-font-smoothing": "0.1.0",
+    "postcss-loader": "3.0.0",
     "query-string": "3.0.0",
-    "resolve-url-loader": "^2.1.0",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "resolve-url-loader": "3.1.2",
     "rimraf": "2.5.2",
-    "rollup": "^0.34.10",
-    "rollup-plugin-buble": "^0.15.0",
-    "sass-loader": "^6.0.6",
+    "rollup": "2.47.0",
+    "sass-loader": "12.1.0",
     "scrolltofixed": "1.0.6",
     "smoothscroll-polyfill": "0.3.4",
-    "svg-fill-loader": "^0.0.8",
-    "url-loader": "0.5.7",
+    "svg-transform-loader": "2.0.13",
+    "svgo": "1.3.2",
+    "svgo-loader": "2.2.1",
+    "tslib": "2.2.0",
+    "url-loader": "^4.0.0",
     "url-search-params": "0.6.1",
-    "webpack": "^3.5.4",
+    "webpack": "5.40.0",
+    "webpack-cli": "4.7.2",
     "whatwg-fetch": "2.0.2"
   },
   "devDependencies": {
     "glob": "7.1.1",
-    "minimist": "1.2.0",
-    "webpack-dev-server": "^2.7.1",
-    "webpack-livereload-plugin": "0.4.0"
+    "minimist": "1.2.3",
+    "webpack-dev-server": "4.0.0-beta.3"
   },
   "scripts": {
-    "build": "webpack --bail",
+    "build": "webpack",
+    "build:production": "NODE_ENV=production webpack  --mode production",
     "postinstall": "cd node_modules/codemirror && rollup -c",
-    "start": "webpack-dev-server"
+    "start": "webpack serve"
   }
 }
