@@ -1,6 +1,6 @@
-module.exports = ({ file, options, env }) => ({
-  sourceMap: true,
-  plugins: {
-    'autoprefixer': { browsers: ['last 2 versions'] },
-  }
-});
+module.exports = {
+  plugins: [
+    require('autoprefixer')({ browsers: ['last 2 versions'] }),
+    require('postcss-custom-properties')
+  ]
+};
