@@ -22,29 +22,29 @@ def process_titles(row_titles, title_prefix, path_folder, suffix):
         process_titles(child_titles, title_prefix, path_folder, suffix)
 
 
-def load_api_titles():
-    api_title_files_path = path.join(root_folder, 'api', 'latest', 'jvm', 'stdlib', 'index.yml')
+def load_api_titles(folder_path):
+    api_title_files_path = path.join(folder_path, 'api', 'latest', 'jvm', 'stdlib', 'index.yml')
     with open(api_title_files_path) as title_files:
         process_titles(yaml.load(title_files)[0], 'latest/jvm/stdlib', '.', '')
 
-    test_title_files_path = path.join(root_folder, 'api', 'latest', 'kotlin.test', 'index.yml')
+    test_title_files_path = path.join(folder_path, 'api', 'latest', 'kotlin.test', 'index.yml')
     with open(test_title_files_path) as title_files:
         process_titles(yaml.load(title_files)[0], 'latest/kotlin.test',  '.', '')
 
 
-def get_api_page(build_mode: bool, page_path):
+def get_api_page(build_mode: bool, page_path, dir_path=root_folder):
     if not page_path.endswith('.html'):
         page_path += '.html'
     if len(titles) == 0:
         try:
-            load_api_titles()
+            load_api_titles(dir_path)
         except FileNotFoundError as e:
             if build_mode:
                 raise e
             else:
                 print("API module is not included: ", e)
 
-    file_path = path.join(root_folder, 'api', page_path)
+    file_path = path.join(dir_path, 'api', page_path)
     if not path.exists(file_path):
         return None
     with open(file_path) as html_file:
