@@ -1,5 +1,7 @@
 import re
 
+replace_simple_code = False
+
 languageMimeTypeMap = {
     "kotlin": "text/x-kotlin",
     "java": "text/x-java",
@@ -19,6 +21,11 @@ languageMimeTypeMap = {
 }
 
 
+def set_replace_simple_code(v: bool):
+    global replace_simple_code
+    replace_simple_code = v
+
+
 def find_closest_tag(element, tagname):
     current_element = element.parent
     while current_element is not None and current_element.name != tagname:
@@ -26,9 +33,53 @@ def find_closest_tag(element, tagname):
     return current_element
 
 
+processors = {
+    'h1': 'typo-header typo-h1',
+    'h2': 'typo-header typo-h2',
+    'h3': 'typo-header typo-h3',
+    'h4': 'typo-header typo-h4',
+
+    'ul': 'typo-list typo-list_type_simple',
+    'ol': 'typo-list typo-list_type_ordered',
+    'li': 'typo-list__item',
+
+    'p': 'typo-para',
+    'a': 'typo-link',
+    'blockquote': 'typo-quote',
+    'hr': 'typo-hr',
+    'img': 'typo-image',
+    'strong': 'typo-strong',
+
+    'table': 'typo-table',
+    'tr': 'typo-table__row',
+    'td': 'typo-table__column',
+}
+
+
+def process_markdown_html(tree):
+    tree = process_code_blocks(tree)
+
+    for element in tree.select('*'):
+        appendClass = processors.get(element.name)
+        if appendClass is not None:
+            if element.has_attr('class'):
+                element['class'].append(processors.get(element.name))
+            else:
+                element['class'] = processors.get(element.name)
+
+    return tree
+
 def process_code_blocks(tree):
-    code_elements = tree.select('pre > code')
-    for element in code_elements:
+    if replace_simple_code:
+        # some spellcheckers may not know what to do with <code> elements,
+        # we replace in-line code blocks with span to improve spellcheckers
+        # TODO: avoid global variable hack here and pass the parameter explicitly
+        for element in tree.select("code"):
+            if len(element.attrs) == 0:
+                element.name = "span"
+                element['style'] = "font-style: italic; text-decoration: underline;"
+
+    for element in tree.select('pre > code'):
         class_names = element.get("class")
         lang = None
         if class_names is not None:
