@@ -1,6 +1,5 @@
 export default {
   all: 'All',
   upcoming: 'Upcoming',
-  past: 'Past',
-  kotlin: 'Kotlin 1.1 Event'
+  past: 'Past'
 };
